requirements.txt
numpy==1.24.3
opencv_python==4.8.0.74
Pillow==9.5.0
Pillow==10.0.0
streamlit==1.25.0
tensorflow==2.13.0
tensorflow_macos==2.13.0
